Started implementing transponder receiver logic
Started implementing transponder receiver logic. Made some changes to
data processor. Everything is now in nautical miles and not some weird
mixture of NM and KM, this may require test changes. Until the mock
transponder works I’m not gonna be able to see if what I’ve done with
the transponder receiver is working correctly……

Not going to lie going this route made things a lot more complicated
and I’m a little concerned about having something that actually works
with in the week instead of just having the mock transponder send the
data in reliable intervals and ignoring the transponder receiver.