Data processor/testing changes
I’ve reverted transponder receiver back to the way it was and now
everything works as it should the way we discussed in class. I agree
that we need to implement the multithreaded listener and have the mock
transponder send individual data from each plane instead of a list and
have the receiver handle putting everything into a list but I don’t
think its a high priority at this time so as of now we can rely on the
mock transponder sending consistent data that is formatted correctly.
If someone does want to implement the transponder receiver
multithreaded functionality feel free to do so but if you do don’t
commit it unless its finished and working.

Anyways I did manage to get it working from the console. I also added a
test to MockTransponderTests so that we can debug and step through the
entire system sendDataTestMac (I just changed the ‘\’ to ‘/‘ to work
for mac).

It is clear that the intersection calculation isn’t working but I did
manually (and painfully) create test data where it does detect a
collision but I think the time until collision is off and not
consistent. It does make it to the audio handler successfully so until
we can find an equation that works you can use this test to test the
audio handler. You can run this test under DataProcessorTests.